Rotate, Zoom and center the image
Improves the rotation by centering the image on the screen when it is
smaller than the screen size.

Optimal zoom is calculated for images than are higher in hight when
they are rotated as well.

Fixed rotation (anti)clockwise. X, Y axis + angles in gradians are
handled different by cairo.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Signed-off-by: Agustin Zubiaga <aguz@sugarlabs.org>