Fix collaboration - SL #4520
Collaboration was broken partially by the remove of __set_file_idle_cb
in 17acb8ec0e72ead145c3d2648707af87b8ff5cf5 and posterior changes.
Now restored the method, but use as a parameter the object_id instead
of the tempfile path, because this change after the file is saved in the
journal.
We also needed add a control in read_file due to the use of the
emptypanel.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>