Enable collaboration only when a file is loaded
Right now, ImageViewer activity can start without a file loaded,
showing the empty panel and a button to open the object chooser.
If the user try sharing the session before a book is opened,
the collaboration is broken, because no file is available to share.
This patch disable collaboration until a file is opened.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>
Signed-off-by: Agustin Zubiaga <aguz@sugarlabs.org>